# docs/swagger/swagger.json
{"openapi":"3.0.1","info":{"title":"XRT Training","version":"1.0.0"},"basePath":"/","components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"bearerAuth":[]}],"paths":{"/":{"get":{"description":"Default route","responses":{"200":{"description":"Success"}}}},"/auth":{"get":{"description":"Register a new user","tags":["Auth"],"responses":{"200":{"description":"Success"}}}},"/auth/register":{"post":{"description":"Register a new user","tags":["Auth"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"test123"},"password":{"type":"string","example":"password123"},"firstname":{"type":"string","example":"john"},"lastname":{"type":"string","example":"smith"},"email":{"type":"string","example":"test@test.com"},"staffid":{"type":"string","example":"abc123"},"isSupervisor":{"type":"boolean","example":true}}}}}},"responses":{"200":{"description":"Success"}}}},"/auth/login":{"post":{"description":"Login","tags":["Auth"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"test123"},"password":{"type":"string","example":"password123"}}}}}},"responses":{"200":{"description":"Success"}}}},"/course/progress":{"post":{"description":"Post a progress update","tags":["Course","Progress"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","example":{"any":"thing"}}}}}}},"responses":{"200":{"description":"Success"}}},"get":{"description":"Get all progress updates","tags":["Course","Progress"],"responses":{"200":{"description":"Success"}}}},"/course":{"get":{"description":"Get all the courses","tags":["Course"],"responses":{"200":{"description":"Success"}}},"post":{"description":"Get a course by id","tags":["Course"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Course"},"description":{"type":"string","example":"This is the first of many courses"}}}}}},"responses":{"200":{"description":"Success"}}}},"/course/{courseId}":{"get":{"description":"Get a course by id","tags":["Course"],"parameters":[{"in":"path","name":"courseId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Update a course by id","tags":["Course"],"parameters":[{"in":"path","name":"courseId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Course"},"description":{"type":"string","example":"This is the first of many courses"}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Get a course by id","tags":["Course"],"parameters":[{"in":"path","name":"courseId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/upload":{"post":{"description":"Upload an Image","tags":["User"],"requestBody":{"required":"true,","content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Success"}}}},"/upload/{username}":{"get":{"description":"Get profile image for user","tags":["User"],"parameters":[{"in":"path","name":"username","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/recipe":{"get":{"description":"Get all the recipes","tags":["Recipe"],"responses":{"200":{"description":"Success"}}}},"/recipe/search":{"get":{"description":"Get recipes by category","tags":["Recipe"],"parameters":[{"in":"query","name":"category","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/recipe/{recipeId}":{"get":{"description":"Get a recipe by id","tags":["Recipe"],"parameters":[{"in":"path","name":"recipeId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Update a recipe by id","tags":["Recipe"],"parameters":[{"in":"path","name":"recipeId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Course"},"steps":{"type":"object"},"category":{"type":"string","example":"Milk Tea"}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Delete a recipe","tags":["Recipe"],"parameters":[{"in":"path","name":"recipeId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/recipe/create":{"post":{"description":"Create a recipe","tags":["Recipe"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Course"},"steps":{"type":"object"},"category":{"type":"string","example":"Milk Tea"}}}}}},"responses":{"200":{"description":"Success"}}}},"/step":{"get":{"description":"Get all steps","tags":["Step"],"responses":{"200":{"description":"Success"}}}},"/step/create":{"post":{"description":"Create a step","tags":["Step"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Step"},"description":{"type":"string","example":"Boil water"}}}}}},"responses":{"200":{"description":"Success"}}}},"/step/{stepId}":{"get":{"description":"Get all steps by name","tags":["Step"],"parameters":[{"in":"path","name":"stepId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Update a step by id","tags":["Step"],"parameters":[{"in":"path","name":"stepId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Step"},"description":{"type":"string","example":"Boil water"}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Delete a step","tags":["Step"],"parameters":[{"in":"path","name":"stepId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/user/all":{"get":{"description":"Get all the users","tags":["User"],"responses":{"200":{"description":"Success"}}}},"/user":{"get":{"description":"Get current user","tags":["User"],"responses":{"200":{"description":"Success"}}}},"/user/{userId}":{"get":{"description":"Get a user by id","tags":["User"],"parameters":[{"in":"path","name":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Get a user by id","tags":["User"],"parameters":[{"in":"path","name":"userId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"test123"},"password":{"type":"string","example":"password123"},"newPassword":{"type":"string","example":"password123"},"firstname":{"type":"string","example":"john"},"lastname":{"type":"string","example":"smith"},"email":{"type":"string","example":"test@test.com"},"staffid":{"type":"string","example":"abc123"},"isSupervisor":{"type":"boolean","example":true}}}}}},"responses":{"200":{"description":"Success"}}}}},"tags":[]}